requirements.txt
feedparser==6.0.11
google-genai==1.14.0
jinja2==3.1.4
requests==2.32.3
python-dotenv==1.0.1
